ResetPasswordPage is now a container and is opened through landing page
