+++ Collibri/ClientApp/src/AppRoutes.js
@@ -2,7 +2,6 @@
 import {Navigate} from "react-router-dom";
 import {LandingPageLayout} from "./components/Layouts/LandingPageLayout";
 import RoomLayout from "./components/Layouts/RoomLayout";
-import ResetPasswordPage from "./components/ResetPasswordPage";
 import AboutPage from "./components/AboutPage";
 
 
@@ -25,7 +24,7 @@ const AppRoutes = [
   //},
   {
     path : '/reset-password/:token',
-    element: <ResetPasswordPage />
+    element: <LandingPageLayout />
   },
   {
     path: '/about',
